revise python sorting codes to python3

+++ WW2/Sorting_Algorithms_Python/heapsort.py
@@ -8,7 +8,7 @@ def swap(array, i, j):
 def heapify(array):
     """ Build heap """
     # Middle in array
-    start = (len(array) - 2) / 2
+    start = int((len(array) - 2) / 2)
     while start >= 0:
         perc_down(array, start, len(array) - 1)
         start -= 1
